backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
pinecone-client
openai
dotenv